requirements.txt
pandas==1.1.4
matplotlib==3.3.2
scipy==1.5.4
torchvision==0.8.1
